Former-commit-id: 3a99fa065977abcf15d3ab6bf31317bf8cd876c4

+++ app/services/address_lookup/gds/WebServiceImpl.scala
@@ -9,7 +9,7 @@ import services.address_lookup.AddressLookupWebService
 final class WebServiceImpl extends AddressLookupWebService {
   private val baseUrl: String = Config.gdsAddressLookupBaseUrl
   private val authorisation: String = Config.gdsAddressLookupAuthorisation
-  private val requestTimeout: Int = Config.gdsAddressLookupRequestTimeout.toInt
+  private val requestTimeout: Int = Config.gdsAddressLookupRequestTimeout
 
   def postcodeWithNoSpaces(postcode: String): String = postcode.filter(_ != ' ')
